Isolate tree-shaken search code from the shipped modules
