+++ src/SearchStrategies/search/findWildcardMatches.ts
@@ -1,7 +1,9 @@
 import { MatchInfo, WildcardConfig } from '../types';
 import { memoizeLast } from '../../utils';
+import { buildWildcardFragment } from './buildWildcardFragment';
 
-/** Memoizes the compiled regex for the last (pattern, maxSpaces), reused across every item in a search. */
+/** Unused — the pre-unified wildcard matcher, tree-shaken from the bundle. See `tests/performance/README.md`. */
+
 const buildWildcardRegex = memoizeLast(
   (pattern: string, config: WildcardConfig) =>
     new RegExp(pattern.replace(/\*/g, buildWildcardFragment(config)), 'gi'),
@@ -14,19 +16,7 @@ function getWildcardRegex(pattern: string, config: WildcardConfig): RegExp {
   return regex;
 }
 
-/**
- * Finds matches using wildcard patterns (* matches any non-space characters).
- * Uses regex to find all matching patterns in the text.
- * Wildcards stop at spaces by default - configure `maxSpaces` to span across words.
- *
- * Unused at runtime (only {@link buildWildcardFragment} is); tree-shaken from the bundle.
- * See `tests/performance/README.md`.
- *
- * @param text - The text to search in
- * @param pattern - The wildcard pattern (e.g., "hel*" matches "hello" but not "hello world")
- * @param config - Optional wildcard configuration
- * @returns Array of MatchInfo objects for each wildcard match
- */
+/** Finds wildcard matches; `*` spans non-space characters, `maxSpaces` lets it cross words. */
 export function findWildcardMatches(text: string, pattern: string, config: WildcardConfig = {}): MatchInfo[] {
   const regex = getWildcardRegex(pattern, config);
   const matches: MatchInfo[] = [];
@@ -46,29 +36,4 @@ export function findWildcardMatches(text: string, pattern: string, config: Wildc
   }
 
   return matches;
-}
-
-export function buildWildcardFragment(config: WildcardConfig): string {
-  const maxSpaces = normalizeMaxSpaces(config.maxSpaces);
-  if (maxSpaces === 0) {
-    return '[^ ]*';
-  }
-
-  if (maxSpaces === Infinity) {
-    return '[^ ]*(?: [^ ]*)*';
-  }
-
-  return `[^ ]*(?: [^ ]*){0,${maxSpaces}}`;
-}
-
-function normalizeMaxSpaces(value: number | undefined): number {
-  if (typeof value !== 'number' || Number.isNaN(value) || value <= 0) {
-    return 0;
-  }
-
-  if (!Number.isFinite(value)) {
-    return Infinity;
-  }
-
-  return Math.floor(value);
 }
