+++ tests/SearchStrategies/findWildcardMatches.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { findWildcardMatches, buildWildcardFragment } from '../../src/SearchStrategies/search/findWildcardMatches';
+import { findWildcardMatches } from '../../src/SearchStrategies/search/findWildcardMatches';
 
 describe('findWildcardMatches', () => {
   it('should return matches for exact matches', () => {
@@ -83,29 +83,6 @@ describe('findWildcardMatches', () => {
     const matches = findWildcardMatches('hello this is a long world sequence', 'hel*rld', { maxSpaces: Infinity });
     expect(matches).toHaveLength(1);
     expect(matches[0].text).toBe('hello this is a long world');
-  });
-});
-
-describe('buildWildcardFragment', () => {
-  it('returns single-word pattern by default', () => {
-    expect(buildWildcardFragment({})).toBe('[^ ]*');
-  });
-
-  it('allows configuring finite spaces', () => {
-    expect(buildWildcardFragment({ maxSpaces: 2 })).toBe('[^ ]*(?: [^ ]*){0,2}');
-  });
-
-  it('normalizes values less than or equal to zero back to default', () => {
-    expect(buildWildcardFragment({ maxSpaces: 0 })).toBe('[^ ]*');
-    expect(buildWildcardFragment({ maxSpaces: -5 })).toBe('[^ ]*');
-  });
-
-  it('supports unlimited spaces with Infinity', () => {
-    expect(buildWildcardFragment({ maxSpaces: Infinity })).toBe('[^ ]*(?: [^ ]*)*');
-  });
-
-  it('floors decimal inputs', () => {
-    expect(buildWildcardFragment({ maxSpaces: 2.9 })).toBe('[^ ]*(?: [^ ]*){0,2}');
   });
 });
